server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pydantic-ai-slim[openrouter,tavily]>=1.84.1",
    "python-dotenv>=1.2.2",
    "sympy>=1.14.0",
    "uvicorn[standard]>=0.44.0",
]
